+++ server.js
@@ -1,11 +1,21 @@
 import express from 'express';
 import cors from 'cors';
 import weatherRoutes from './backend/routes/weather.js';
+import path from 'path';
+import { fileURLToPath } from 'url';
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
 
 const app = express();
 
 app.use(express.json());
 app.use(cors());
+
+app.use(express.static(path.join(__dirname, 'frontend')));
+app.get('/', (req, res) => {
+  res.sendFile(path.join(__dirname, 'frontend', 'index.html'));
+});
 
 app.use('/api', weatherRoutes);